pyproject.toml
[project]
name = "pricetracker"
version = "0.1.0"
description = "Product Price Tracker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "colorlog>=6.9.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "schedule>=1.2.2",
]
